wip on group funcs
